requirements.txt
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
tqdm>=4.66.0
streamlit>=1.41.0
openpyxl>=3.1.0
pytest>=8.0.0
